Revamp scalability module
The scalability module was previously implemented with the incorrect
understanding that the return values of `bind` and `connect` were simple
status codes and not endpoint ids that should be used for sending and
receiving. This has been fixed.

There are new procedures to handle socket and endpoint lifecycle
management. New as well as existing procedures have been broken up into
a set of primitives which deal on the plane of result codes, while
higher-level procedures check those result codes, distinguish between
timeouts and other errors, and throw exceptions for the latter.